Make sure to check if ssh-agent is running before starting it
This way ssh-agent will be started only once when terminal is opened. As tmux is
started after the terminal is opened, this ssh-agent will be shared by all tmux
windows and panes. This way ssh-add command only needs to be ran once and ssh
key will be shared between tmux windows and panes.